Android JGit: Update the version
We want to use this specific version number as it is the only one that
supports `git clone` on Android. The others use a hard-link while
creating a lock file while trying to perform a 'git clone'.

The Android Internal Storage does not support hard links, and therefore
aborts. This will need to be fixed in the latest version of jgit.