Rename NoteBrowsingScreen to JournalBrowsingScreen
I'm add a different interface for viewing Notes, similar to card viewer,
and it'll be nice to have a distinction between Journal viewing and Note
Viewing. The biggest difference is that when viewing a Journal, there is
a timeline and you want to be able to easily move between Journal
entries. Whereas with notes, they aren't connected in any obvious way.