Remote Setup: Only generate the key if not already generated
This way one can go back and change the clone url if one wants to and
the ssh key would not have changed. There is an additional SSH Key
changing button if one needs to regenerate it for some reason.

This should fix #47. Well, it's not the best solution, but it is a
workaround.

For the Clone Url you must enter SSH-KEY-ID@clone-url-provided-by-aws.
In order to get the SSH KEY ID you must upload the public key provdied
to your AWS Credentials.