Remove FontSize configuration
Ever since we had a proper markdown renderer, this option has been
broken as it only scales the body text and not the headers, and other
text in code-blocks. Additionally it only modifies the viewing text size
and not the editing text size. Fixing this feature is way too much work,
and barely any users are using this feature.

Maybe we will introduce it again in the future.