Graph Rendering Experiment: Try layouting via a RenderBox
Instead of using a Custom Painter and then re-implementing all the hit
testing and everything, instead lets try to learn about Flutter's lower
level RenderObjects and let each node just be a simple Widget and
provide a GraphView class, which will take a list of nodes (widgets),
and edges.

This is a small PoC, which looks quite promising as instead of drawing a
circle as a node, we will be able to render the node as how a Note is
rendered in Markdown.