Cleanup PurchaseScreen
Avoid having one widget call the other widget with the sku price. It was
complicating the entire code base. Also, create a new class
PurchaseManager which is a singleton and will be used to manage all
purchases from now on.

The earlier approach of having the code in a widget would sometimes
result in purchases not being confirmed. I'm hoping that putting the
class in a singleton will help.