Bump deps for Flutter 2
Also for handling the latest dart-git. This required disabling -
* ssh_key - Feature toggle (off by default) for generating the ssh key
in dart. This was causing problems.
* screenshots
* flutter_driver - used by screenshots

flutter_driver was nice to have as it was the only proper integration
test that we have.

Also, many of the tests currently fail because of problems with Sentry.
I prefer to fix those in another commit